requirements.txt
fastapi
torch
transformers
pillow
uvicorn